Mark readme and changelog as disabled if they're not reachable
We can tell that the readme and changelog won't be reachable if
we're unable to find a Github or Gitlab repo using the available
URLs in the gem metadata. In that case, we now disable those
options in the TUI to prevent confusion.

It's still possible that the readme and changelog don't exist
but it is less likely.