Remove the dry-struct gem
This was pulling in a bunch of other dependencies which was not
great and it wasn't too difficult to remove (though we get less
type-safety). I also changed out the snapshot testing library
to one that allows for better support for custom formatters
so that I can use the YAML gem as a formatter. This handles
pretty printing internal values so special introspection
methods and libraries don't need to get used instead which is
nice.